remove UA handler
